+++ src/lets-version.js
@@ -8,9 +8,11 @@
  * @typedef {import('type-fest').PackageJson} PackageJson
  * @typedef {import('./changelog.js').GenerateChangelogOpts} GenerateChangelogOpts
  * @typedef {import('./dependencies.js').SynchronizeBumpsReturnType} SynchronizeBumpsReturnType
+ * @typedef {import('./readUserConfig.js').LetsVersionConfig} LetsVersionConfig
  */
 
 import appRootPath from 'app-root-path';
+import { execSync } from 'child_process';
 import { detect as detectPM } from 'detect-package-manager';
 import fs from 'fs-extra';
 import os from 'os';
@@ -21,7 +23,6 @@ import semver from 'semver';
 import { getChangelogUpdateForPackageInfo, getFormattedChangelogDate } from './changelog.js';
 import { fixCWD } from './cwd.js';
 import { getBumpRecommendationForPackageInfo, synchronizeBumps } from './dependencies.js';
-import { execAsync } from './exec.js';
 import { filterPackagesByNames, getAllPackagesChangedBasedOnFilesModified, getPackages } from './getPackages.js';
 import {
   formatVersionTagForPackage,
@@ -35,9 +36,11 @@ import {
   gitWorkdirUnclean,
 } from './git.js';
 import { conventionalCommitToBumpType } from './parser.js';
+import { defineLetsVersionConfig, readLetsVersionConfig } from './readUserConfig.js';
 import {
   BumpType,
   BumpTypeToString,
+  ChangelogAggregateUpdate,
   ChangelogEntryType,
   ChangelogUpdate,
   ChangelogUpdateEntry,
@@ -46,29 +49,42 @@ import {
   ReleaseAsPresets,
 } from './types.js';
 
+export { defineLetsVersionConfig };
+
+/**
+ * @typedef {Object} AllCommandsBaseOpts
+ * @property {string} [cwd=appRootPath.toString()]
+ */
+
 /**
  * Returns all detected packages for this repository
  *
- * @param {string} [cwd=appRootPath.toString()]
+ * @param {AllCommandsBaseOpts} [opts]
  * @returns {Promise<PackageInfo[]>}
  */
-export async function listPackages(cwd = appRootPath.toString()) {
-  const fixedCWD = fixCWD(cwd);
+export async function listPackages(opts) {
+  const fixedCWD = fixCWD(opts?.cwd || appRootPath.toString());
 
   return filterPackagesByNames(await getPackages(fixedCWD), undefined, fixedCWD);
 }
 
 /**
+ * @typedef {Object} GetLastVersionTagsByPackageNameOpts
+ * @property {string} [cwd=appRootPath.toString()]
+ * @property {string[]} [names]
+ * @property {boolean} [noFetchTags=false]
+ */
+
+/**
  * Given an optional array of package names, reads the latest
  * git tag that was used in a previous version bump operation.
  *
- * @param {string[]} [names]
- * @param {boolean} [noFetchTags=false]
- * @param {string} [cwd=appRootPath.toString()]
+ * @param {GetLastVersionTagsByPackageNameOpts} [opts]
  *
  * @returns {Promise<PublishTagInfo[]>}
  */
-export async function getLastVersionTagsByPackageName(names, noFetchTags = false, cwd = appRootPath.toString()) {
+export async function getLastVersionTagsByPackageName(opts) {
+  const { names, noFetchTags = false, cwd = appRootPath.toString() } = opts ?? {};
   const fixedCWD = fixCWD(cwd);
 
   const filteredPackages = await filterPackagesByNames(await getPackages(fixedCWD), names, fixedCWD);
@@ -82,12 +98,11 @@ export async function getLastVersionTagsByPackageName(names, noFetchTags = false
  * Gets a list of all files that have changed since the last publish for a specific package or set of packages.
  * If no results are returned, it likely means that there was not a previous version tag detected in git.
  *
- * @param {string[]} [names]
- * @param {boolean} [noFetchTags=false]
- * @param {string} [cwd=appRootPath.toString()]
+ * @param {GetLastVersionTagsByPackageNameOpts} [opts]
  * @returns {Promise<string[]>}
  */
-export async function getChangedFilesSinceBump(names, noFetchTags = false, cwd = appRootPath.toString()) {
+export async function getChangedFilesSinceBump(opts) {
+  const { names, noFetchTags = false, cwd = appRootPath.toString() } = opts ?? {};
   const fixedCWD = fixCWD(cwd);
   const filteredPackages = await filterPackagesByNames(await getPackages(fixedCWD), names, fixedCWD);
 
@@ -101,13 +116,13 @@ export async function getChangedFilesSinceBump(names, noFetchTags = false, cwd =
 /**
  * Gets a list of all packages that have changed since the last publish for a specific package or set of packages.
  * If no results are returned, it likely means that there was not a previous version tag detected in git.
- * @param {string[]} [names]
- * @param {boolean} [noFetchTags]
- * @param {string} [cwd=appRootPath.toString()]
+ *
+ * @param {GetLastVersionTagsByPackageNameOpts} [opts]
  *
  * @returns {Promise<PackageInfo[]>}
  */
-export async function getChangedPackagesSinceBump(names, noFetchTags = false, cwd = appRootPath.toString()) {
+export async function getChangedPackagesSinceBump(opts) {
+  const { names, noFetchTags = false, cwd = appRootPath.toString() } = opts ?? {};
   const fixedCWD = fixCWD(cwd);
 
   const filteredPackages = await filterPackagesByNames(await getPackages(fixedCWD), names, fixedCWD);
@@ -121,16 +136,23 @@ export async function getChangedPackagesSinceBump(names, noFetchTags = false, cw
 }
 
 /**
+ * @typedef {Object} GetConventionalCommitsByPackageOpts
+ * @property {string} [cwd=appRootPath.toString()]
+ * @property {string[]} [names]
+ * @property {boolean} [noFetchTags=false]
+ * @property {boolean} [noFetchAll=false]
+ */
+
+/**
  * Parses commits since last publish for a specific package or set of packages
  * and returns them represented as Conventional Commits objects.
  *
- * @param {string[]} [names]
- * @param {boolean} [noFetchAll=false]
- * @param {string} [cwd=appRootPath.toString()]
+ * @param {GetConventionalCommitsByPackageOpts} [opts]
  *
  * @returns {Promise<GitCommitWithConventionalAndPackageInfo[]>}
  */
-export async function getConventionalCommitsByPackage(names, noFetchAll = false, cwd = appRootPath.toString()) {
+export async function getConventionalCommitsByPackage(opts) {
+  const { names, noFetchAll = false, cwd = appRootPath.toString() } = opts ?? {};
   const fixedCWD = fixCWD(cwd);
 
   const filteredPackages = await filterPackagesByNames(await getPackages(fixedCWD), names, fixedCWD);
@@ -149,37 +171,45 @@ export async function getConventionalCommitsByPackage(names, noFetchAll = false,
  */
 
 /**
+ * @typedef {Object} GetRecommendedBumpsByPackageOpts
+ * @property {string[]} [names]
+ * @property {ReleaseAsPresets} [releaseAs='auto']
+ * @property {string} [preid='']
+ * @property {boolean} [uniqify=false]
+ * @property {boolean} [forceAll=false]
+ * @property {boolean} [noFetchAll=false]
+ * @property {boolean} [noFetchTags=false]
+ * @property {boolean} [updatePeer=false]
+ * @property {boolean} [updateOptional=false]
+ * @property {string} [cwd=appRootPath.toString()]
+ */
+
+/**
  * Given an optional list of package names, parses the git history
  * since the last bump operation and suggests a bump.
  *
  * NOTE: It is possible for your bump recommendation to not change.
  * If this is the case, this means that your particular package has never had a version bump by the lets-version library.
  *
- * @param {string[]} [names]
- * @param {ReleaseAsPresets} [releaseAs='auto']
- * @param {string} [preid='']
- * @param {boolean} [uniqify=false]
- * @param {boolean} [forceAll=false]
- * @param {boolean} [noFetchAll=false]
- * @param {boolean} [noFetchTags=false]
- * @param {boolean} [updatePeer=false]
- * @param {boolean} [updateOptional=false]
- * @param {string} [cwd=appRootPath.toString()]
+ * @param {GetRecommendedBumpsByPackageOpts} [opts]
  *
  * @returns {Promise<GetRecommendedBumpsByPackageReturnType>}
  */
-export async function getRecommendedBumpsByPackage(
-  names,
-  releaseAs = ReleaseAsPresets.AUTO,
-  preid = '',
-  uniqify = false,
-  forceAll = false,
-  noFetchAll = false,
-  noFetchTags = false,
-  updatePeer = false,
-  updateOptional = false,
-  cwd = appRootPath.toString(),
-) {
+export async function getRecommendedBumpsByPackage(opts) {
+  const {
+    names,
+    releaseAs = ReleaseAsPresets.AUTO,
+    uniqify = false,
+    forceAll = false,
+    noFetchAll = false,
+    noFetchTags = false,
+    updatePeer = false,
+    updateOptional = false,
+    cwd = appRootPath.toString(),
+  } = opts ?? {};
+
+  let preid = opts?.preid || '';
+
   /**
    * @type {GenerateChangelogOpts}
    */
@@ -323,6 +353,26 @@ export async function getRecommendedBumpsByPackage(
 }
 
 /**
+ * @typedef {Object} ApplyRecommendedBumpsByPackageOpts
+ * @property {string[]} [names]
+ * @property {ReleaseAsPresets} [releaseAs='auto']
+ * @property {string} [preid='']
+ * @property {boolean} [uniqify=false]
+ * @property {boolean} [forceAll=false]
+ * @property {boolean} [noFetchAll=false]
+ * @property {boolean} [noFetchTags=false]
+ * @property {boolean} [yes=false] - If true, skips all user confirmations
+ * @property {boolean} [updatePeer=false] - If true, will update any dependent "package.json#peerDependencies" fields
+ * @property {boolean} [updateOptional=false] - If true, will update any dependent "package.json#optionalDependencies" fields
+ * @property {boolean} [noPush=false] - If true, will prevent pushing any changes to upstream / origin
+ * @property {boolean} [rollupChangelog=false] - If true, in addition to updating changelog files for all packages that will be bumped, creates a "rollup" CHANGELOG.md at the root of the repo that contains an aggregate of changes
+ * @property {boolean} [noChangelog=false] - If true, will not write CHANGELOG.md updates for each package that has changed
+ * @property {boolean} [dryRun=false] - If true, will print the changes that are expected to happen at every step instead of actually writing the changes
+ * @property {string} [cwd=appRootPath.toString()]
+ * @property {LetsVersionConfig} [customConfig]
+ */
+
+/**
  * Given an optional list of package names, parses the git history
  * since the last bump operation, suggest a bump and applies it, also
  * updating any dependent package.json files across your repository.
@@ -330,45 +380,33 @@ export async function getRecommendedBumpsByPackage(
  * NOTE: It is possible for your bump recommendation to not change.
  * If this is the case, this means that your particular package has never had a version bump by the lets-version library.
  *
- * @param {string[]} [names]
- * @param {ReleaseAsPresets} [releaseAs='auto']
- * @param {string} [preid='']
- * @param {boolean} [uniqify=false]
- * @param {boolean} [forceAll=false]
- * @param {boolean} [noFetchAll=false]
- * @param {boolean} [noFetchTags=false]
- * @param {object} [opts]
- * @param {boolean} [opts.yes=false] - If true, skips all user confirmations
- * @param {boolean} [opts.updatePeer=false] - If true, will update any dependent "package.json#peerDependencies" fields
- * @param {boolean} [opts.updateOptional=false] - If true, will update any dependent "package.json#optionalDependencies" fields
- * @param {boolean} [opts.noPush=false] - If true, will prevent pushing any changes to upstream / origin
- * @param {boolean} [opts.noChangelog=false] - If true, will not write CHANGELOG.md updates for each package that has changed
- * @param {boolean} [opts.dryRun=false] - If true, will print the changes that are expected to happen at every step instead of actually writing the changes
- * @param {ChangeLogLineFormatter} [opts.changelogLineFormatter] - If provided, will be used to format the changelog line for each package that has changed
- * @param {ChangeLogEntryFormatter} [opts.changelogEntryFormatter] - If provided, will be used to format the changelog entry for all of the changes for a version
- * @param {string} [cwd=appRootPath.toString()]
+ * @param {ApplyRecommendedBumpsByPackageOpts} [opts]
  *
  * @returns {Promise<GetRecommendedBumpsByPackageReturnType | null>}
  */
-export async function applyRecommendedBumpsByPackage(
-  names,
-  releaseAs = ReleaseAsPresets.AUTO,
-  preid = '',
-  uniqify = false,
-  forceAll = false,
-  noFetchAll = false,
-  noFetchTags = false,
-  opts,
-  cwd = appRootPath.toString(),
-) {
+export async function applyRecommendedBumpsByPackage(opts) {
+  const {
+    customConfig: customConfigOverride,
+    cwd = appRootPath.toString(),
+    dryRun = false,
+    forceAll = false,
+    names,
+    noChangelog = false,
+    noFetchAll = false,
+    noFetchTags = false,
+    noPush = false,
+    preid = '',
+    releaseAs = ReleaseAsPresets.AUTO,
+    rollupChangelog = false,
+    uniqify = false,
+    updateOptional = false,
+    updatePeer = false,
+  } = opts ?? {};
   const fixedCWD = fixCWD(cwd);
 
+  const customConfig = customConfigOverride ?? (await readLetsVersionConfig(fixedCWD));
+
   let yes = opts?.yes || false;
-  const updatePeer = opts?.updatePeer || false;
-  const updateOptional = opts?.updateOptional || false;
-  const noPush = opts?.noPush || false;
-  const noChangelog = opts?.noChangelog || false;
-  const dryRun = opts?.dryRun || false;
 
   if (dryRun) console.warn('**Dry Run has been enabled**');
 
@@ -383,7 +421,8 @@ export async function applyRecommendedBumpsByPackage(
 
   if (!allPackages.length) return null;
 
-  const synchronized = await getRecommendedBumpsByPackage(
+  const synchronized = await getRecommendedBumpsByPackage({
+    cwd: fixedCWD,
     names,
     releaseAs,
     preid,
@@ -393,8 +432,7 @@ export async function applyRecommendedBumpsByPackage(
     noFetchTags,
     updatePeer,
     updateOptional,
-    fixedCWD,
-  );
+  });
   const { bumpsByPackageName: presyncBumpsByPackageName } = synchronized;
 
   if (!synchronized.bumps.length) {
@@ -436,8 +474,11 @@ export async function applyRecommendedBumpsByPackage(
   await Promise.all(
     synchronized.bumps.map(async b => {
       if (dryRun) {
-        console.info(`Will write package.json updates for ${b.packageInfo.name} to ${b.packageInfo.packageJSONPath}`);
-      } else fs.writeFile(b.packageInfo.packageJSONPath, JSON.stringify(b.packageInfo.pkg, null, 2), 'utf-8');
+        return console.info(
+          `Will write package.json updates for ${b.packageInfo.name} to ${b.packageInfo.packageJSONPath}`,
+        );
+      }
+      return fs.writeFile(b.packageInfo.packageJSONPath, JSON.stringify(b.packageInfo.pkg, null, 2), 'utf-8');
     }),
   );
 
@@ -455,16 +496,17 @@ export async function applyRecommendedBumpsByPackage(
      * and
      * https://github.com/npm/cli/issues/4942
      */
-    const syncLockfiles = async () => {
+    const syncLockfiles = () => {
+      if (didSyncLockFiles) return;
       try {
-        await execAsync(`${pm} install`, { cwd: fixedCWD, stdio: 'ignore' });
+        execSync(`${pm} install`, { cwd: fixedCWD, stdio: 'inherit' });
         didSyncLockFiles = true;
       } catch (error) {
         didSyncLockFiles = false;
       }
     };
-    await syncLockfiles();
-    await syncLockfiles();
+    syncLockfiles();
+    syncLockfiles();
 
     if (!didSyncLockFiles) {
       console.error('Failed to synchronize lock files. Aborting remaining operations');
@@ -480,7 +522,7 @@ export async function applyRecommendedBumpsByPackage(
     const changelogInfo = await getChangelogUpdateForPackageInfo({
       commits: synchronized.conventional,
       bumps: synchronized.bumps,
-      lineFormatter: opts?.changelogLineFormatter,
+      lineFormatter: customConfig?.changelog?.changelogLineFormatter,
     });
 
     for (const syncbump of synchronized.bumps) {
@@ -502,7 +544,7 @@ export async function applyRecommendedBumpsByPackage(
                 sha: '',
               }),
             ],
-            opts?.changelogLineFormatter,
+            customConfig?.changelog?.changelogLineFormatter,
           ),
         }),
       );
@@ -521,9 +563,8 @@ export async function applyRecommendedBumpsByPackage(
         } catch (error) {
           /* file doesn't exist */
         }
-        const changelogUpdates = opts?.changelogEntryFormatter
-          ? opts.changelogEntryFormatter(c)
-          : `${c.toString()}${os.EOL}---${os.EOL}${os.EOL}`;
+        const changelogUpdates =
+          customConfig?.changelog?.changeLogEntryFormatter?.(c) ?? `${c.toString()}${os.EOL}---${os.EOL}${os.EOL}`;
 
         if (dryRun) {
           console.info(
@@ -532,6 +573,39 @@ export async function applyRecommendedBumpsByPackage(
         } else await fs.writeFile(c.changelogPath, `${changelogUpdates}${existingChangelog}`, 'utf-8');
       }),
     );
+
+    if (rollupChangelog) {
+      // User wants an aggregated changelog at the root.
+      // if this repo only has a single package AND that package is marked
+      // as the root package, do nothing
+
+      const continueWithRollupChangelog =
+        (synchronized.packages.length === 1 && !synchronized.packages[0]?.root) || synchronized.packages.length > 1;
+
+      if (continueWithRollupChangelog) {
+        const changelogUpdates = new ChangelogAggregateUpdate(fixedCWD, getFormattedChangelogDate(), changelogInfo);
+        if (!dryRun) await fs.ensureFile(changelogUpdates.changelogPath);
+
+        let existingChangelog = '';
+        try {
+          existingChangelog = await fs.readFile(changelogUpdates.changelogPath, 'utf-8');
+        } catch (error) {}
+
+        const updatesToWrite = customConfig?.changelog?.changeLogRollupFormatter
+          ? customConfig?.changelog?.changeLogRollupFormatter(changelogUpdates)
+          : changelogUpdates.toString();
+
+        if (dryRun) {
+          console.info(
+            `Will write the following rollup changelog updated to ${changelogUpdates.changelogPath}:${os.EOL}${os.EOL}${
+              updatesToWrite || ''
+            }`,
+          );
+        } else if (updatesToWrite) {
+          await fs.writeFile(changelogUpdates.changelogPath, `${updatesToWrite}${existingChangelog}`, 'utf-8');
+        }
+      }
+    }
   }
 
   // commit the stuffs